+++ src/linear_model/mod.rs
@@ -3,6 +3,8 @@
 mod least_squares;
 mod linear;
 mod logistic;
+mod ridge;
 
 pub use linear::{LinearRegression, LinearRegressionParams};
 pub use logistic::{LogisticRegression, LogisticRegressionParams};
+pub use ridge::{Ridge, RidgeParams};